Splitting out navigation logic into a common class
