+++ admin/src/Membership/GroupList.jsx
@@ -4,6 +4,7 @@ import Collection from "../Models/Collection";
 import CollectionTable from "../Components/CollectionTable";
 import Group from "../Models/Group";
 import SearchBox from "../Components/SearchBox";
+import CollectionNavigation from "../Models/CollectionNavigation";
 
 const Row = props => {
     const {item, deleteItem} = props;
@@ -19,38 +20,14 @@ const Row = props => {
 };
 
 
-class GroupList extends React.Component {
+class GroupList extends CollectionNavigation {
 
     constructor(props) {
         super(props);
-        this.onSearch = this.onSearch.bind(this);
-        this.onPageNav = this.onPageNav.bind(this);
+        const {search_term, page} = this.state;
 
-        this.params = new URLSearchParams(this.props.location.search);
-        const search_term = this.params.get('search') || '';
-        const page = this.params.get('page') || 1;
         this.collection = new Collection({type: Group, search: search_term, page: page});
-        this.state = {'search': search_term, 'page_index': page};
     }
-
-    onSearch(term) {
-        this.setState({'search': term});
-        this.collection.updateSearch(term);
-        if (term === "") {
-            this.params.delete("search");
-        } else {
-            this.params.set("search", term);
-        }
-        this.props.history.replace(this.props.location.pathname + "?" + this.params.toString());
-    }
-
-    onPageNav(index) {
-        this.setState({'page_index': index});
-        this.collection.updatePage(index);
-        this.params.set("page", index);
-        this.props.history.replace(this.props.location.pathname + "?" + this.params.toString());
-    }
-
 
     render() {
         const columns = [
