+++ admin/src/Membership/MemberList.jsx
@@ -5,6 +5,7 @@ import Collection from "../Models/Collection";
 import Member from "../Models/Member";
 import CollectionTable from "../Components/CollectionTable";
 import SearchBox from "../Components/SearchBox";
+import CollectionNavigation from "../Models/CollectionNavigation";
 
 const Row = props => {
     const {item, deleteItem} = props;
@@ -21,36 +22,14 @@ const Row = props => {
 };
 
 
-class MemberList extends React.Component {
+class MemberList extends CollectionNavigation {
 
     constructor(props) {
         super(props);
-        this.onSearch = this.onSearch.bind(this);
-        this.onPageNav = this.onPageNav.bind(this);
+        const {search_term, page} = this.state;
+        console.log(this.state);
 
-        this.params = new URLSearchParams(this.props.location.search);
-        const search_term = this.params.get('search') || '';
-        const page = this.params.get('page') || 1;
         this.collection = new Collection({type: Member, search: search_term, page: page});
-        this.state = {'search': search_term, 'page_index': page};
-    }
-
-    onSearch(term) {
-        this.setState({'search': term});
-        this.collection.updateSearch(term);
-        if (term === "") {
-            this.params.delete("search");
-        } else {
-            this.params.set("search", term);
-        }
-        this.props.history.replace(this.props.location.pathname + "?" + this.params.toString());
-    }
-
-    onPageNav(index) {
-        this.setState({'page_index': index});
-        this.collection.updatePage(index);
-        this.params.set("page", index);
-        this.props.history.replace(this.props.location.pathname + "?" + this.params.toString());
     }
